working version. Needs more testdata
