+++ src/main/java/com/communityratesgames/datainit/DataInitListener.java
@@ -1,39 +1,31 @@
 package com.communityratesgames.datainit;
 
+import com.communityratesgames.dao.DataAccessLocal;
 import com.communityratesgames.domain.Company;
-import com.communityratesgames.domain.Game;
-import com.communityratesgames.domain.Platform;
 import com.communityratesgames.model.GameModel;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
+import javax.inject.Inject;
 import javax.servlet.ServletContextEvent;
-import java.io.IOException;
 import java.io.InputStream;
 import java.util.*;
 
 public class DataInitListener implements javax.servlet.ServletContextListener {
 
-    @PersistenceContext(unitName = "communityratesgames")
-    private EntityManager em;
+    @Inject
+    private DataAccessLocal dal;
 
     @Override
     public void contextInitialized(ServletContextEvent servletContextEvent) {
-        derp();
+        TestDataInitializer();
     }
 
     @Override
     public void contextDestroyed(ServletContextEvent servletContextEvent) {
 
     }
-    public void derp() {
+    public void TestDataInitializer() {
         ObjectMapper mapper = new ObjectMapper();
-        System.out.println("###################");
-        System.out.println("###################");
-        System.out.println("###################");
-        System.out.println("###################");
-
         try {
             ClassLoader classloader = Thread.currentThread().getContextClassLoader();
             InputStream is = classloader.getResourceAsStream("testdata.json");
@@ -41,44 +33,26 @@ public class DataInitListener implements javax.servlet.ServletContextListener {
             createCompanies(dataHolder.companies);
             createPlatforms(dataHolder.platforms);
             createGames(dataHolder.games);
-        }catch(IOException e) {
-            System.out.println("ERROL" + e);
+        }catch(Exception e) {
         }
     }
     private void createCompanies(List<Company> companies){
-        companies.forEach(c -> em.persist(c));
+        companies.forEach(c -> dal.registerNewCompany(c));
     }
 
     private void createPlatforms(List<Map<String,Object>> platforms) {
-        platforms.forEach(p -> em.persist(new Platform(
-                (String)p.get("name"),
-                (int)p.get("releaseYear"),
-                getCompanyEntity((String) p.get("company")),
-                null
-        )));
+        platforms.forEach(p -> {
+            long l = (int)p.get("company");
+            dal.createPlatform(
+                    (String)p.get("name"),
+                    (int)p.get("releaseYear"),
+                    l
+            );}
+        );
     }
 
     private void createGames(List<GameModel> games){
-            games.forEach(g -> em.persist(createGameEntity(g)));
-    }
-
-    private Game createGameEntity(GameModel model) {
-        return new Game(
-                model.getReleaseDate(),
-                model.getTitle(),
-                getCompanyEntity(model.getCompany()),
-                getPlatformEntity(model.getPlatforms())
-        );
-    }
-    private Company getCompanyEntity(String companyname) {
-        return em.createQuery("SELECT c FROM Company c WHERE c.companyName = :name",Company.class)
-                .setParameter("name", companyname)
-                .getSingleResult();
-    }
-    private List<Platform> getPlatformEntity(List<String> platformList){
-        return em.createQuery("SELECT p FROM Platform p WHERE p.name IN :list",Platform.class)
-                .setParameter("list", platformList)
-                .getResultList();
+            games.forEach(g -> dal.addGame(g));
     }
 
     static class DataCollection {
